# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Incidents, service requests, workflow, assignment, comments and audit for IT service management. Errors are RFC 9457 problem details (`ApiProblem`); clients should branch on `code`.",
    "title" : "Enterprise Service Management API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Same origin as the web app"
  } ],
  "security" : [ {
    "bearer-jwt" : [ ]
  } ],
  "paths" : {
...
    "/api/incidents" : {
      "post" : {
        "operationId" : "createIncident",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateIncidentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketCreatedResponse"
                }
              }
            },
            "description" : "Created"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Report an incident. Priority is calculated and the ticket is routed to a team",
        "tags" : [ "Incidents" ]
      }
    },
    "/api/service-requests" : {
      "post" : {
        "operationId" : "createServiceRequest",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateServiceRequestRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketCreatedResponse"
                }
              }
            },
            "description" : "Created"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Submit a service request; it is routed to a team by category",
        "tags" : [ "Service requests" ]
      }
    },
...
    "/api/tickets" : {
      "get" : {
        "operationId" : "listTickets",
        "parameters" : [ {
          "in" : "query",
          "name" : "view",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "ALL", "REQUESTED", "MINE", "TEAM", "UNASSIGNED" ]
          }
        }, {
          "in" : "query",
          "name" : "type",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "INCIDENT", "SERVICE_REQUEST" ]
            }
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "in" : "query",
          "name" : "priority",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "P1", "P2", "P3", "P4" ]
            }
          }
        }, {
          "in" : "query",
          "name" : "teamId",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "assigneeId",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "categoryId",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "createdFrom",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "createdTo",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "open",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 20,
            "minimum" : 1
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseTicketSummary"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Queues, filters and search over tickets the caller may see (paginated)",
        "tags" : [ "Tickets" ]
      }
    },
...
    "/api/tickets/{id}/transitions" : {
      "get" : {
        "operationId" : "listAvailableTransitions",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AvailableTransition"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Status changes the caller can make now, with labels and required inputs",
        "tags" : [ "Ticket workflow" ]
      },
      "post" : {
        "operationId" : "transitionTicket",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TicketResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Error: see `code` (validation 400, forbidden 403, not found 404, conflict 409, rule violation 422)"
          }
        },
        "summary" : "Move the ticket to another status (requires current version)",
        "tags" : [ "Ticket workflow" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiProblem" : {
        "type" : "object",
        "description" : "RFC 9457 problem details, returned for every error",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Stable machine-readable code: branch on this, never on detail",
            "enum" : [ "VALIDATION_FAILED", "MALFORMED_REQUEST", "AUTHENTICATION_REQUIRED", "ACCESS_DENIED", "RESOURCE_NOT_FOUND", "METHOD_NOT_ALLOWED", "UNSUPPORTED_MEDIA_TYPE", "CONCURRENT_MODIFICATION", "INVALID_SORT_PROPERTY", "INVALID_CREDENTIALS", "REFRESH_TOKEN_INVALID", "INVALID_ORIGIN", "CURRENT_PASSWORD_INCORRECT", "EMAIL_ALREADY_EXISTS", "LAST_ADMIN_REQUIRED", "TEAM_NAME_ALREADY_EXISTS", "TEAM_INACTIVE", "USER_NOT_ELIGIBLE_FOR_TEAM", "CATEGORY_CODE_ALREADY_EXISTS", "CATEGORY_TEAM_REQUIRED", "INVALID_CATEGORY", "INVALID_STATUS_TRANSITION", "TRANSITION_NOT_PERMITTED", "TRANSITION_REQUIREMENT_MISSING", "TICKET_NOT_ASSIGNED", "ASSIGNMENT_NOT_PERMITTED", "ASSIGNEE_NOT_IN_TEAM", "TICKET_NOT_ASSIGNABLE", "MEMBER_HAS_OPEN_TICKETS", "COMMENT_NOT_PERMITTED", "TICKET_CLOSED", "REQUEST_FAILED", "INTERNAL_ERROR" ]
          },
          "correlationId" : {
            "type" : [ "string", "null" ],
            "description" : "Matches the X-Request-Id response header and server log lines"
          },
          "detail" : {
            "type" : [ "string", "null" ],
            "description" : "Human-readable explanation; never contains internal details"
          },
          "fieldErrors" : {
            "type" : [ "array", "null" ],
            "description" : "Present for VALIDATION_FAILED",
            "items" : {
              "$ref" : "#/components/schemas/FieldErrorDetail"
            }
          },
          "instance" : {
            "type" : [ "string", "null" ],
            "description" : "Request path",
            "example" : "/api/tickets/0199b2c4-6a1e-7c3e-9f00-1a2b3c4d5e6f/transitions"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 409
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "title" : {
            "type" : "string",
            "example" : "Conflict"
          },
          "type" : {
            "type" : "string",
            "description" : "Problem type URI",
            "example" : "about:blank"
          }
        },
        "required" : [ "code", "correlationId", "detail", "fieldErrors", "instance", "status", "timestamp", "title", "type" ]
      },
...
      "AvailableTransition" : {
        "type" : "object",
        "properties" : {
          "label" : {
            "type" : "string"
          },
          "requirements" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "REASON", "RESOLUTION", "FULFILMENT_NOTES", "ASSIGNEE" ]
            },
            "uniqueItems" : true
          },
          "targetStatus" : {
            "type" : "string"
          }
        },
        "required" : [ "label", "requirements", "targetStatus" ]
      },
...
      "CreateIncidentRequest" : {
        "type" : "object",
        "properties" : {
          "affectedService" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "categoryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "description" : {
            "type" : "string",
            "maxLength" : 10000,
            "minLength" : 0
          },
          "impact" : {
            "type" : "string",
            "enum" : [ "LOW", "MEDIUM", "HIGH" ]
          },
          "subcategoryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          },
          "urgency" : {
            "type" : "string",
            "enum" : [ "LOW", "MEDIUM", "HIGH" ]
          }
        },
        "required" : [ "categoryId", "description", "impact", "title", "urgency" ]
      },
      "CreateServiceRequestRequest" : {
        "type" : "object",
        "properties" : {
          "categoryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "description" : {
            "type" : "string",
            "maxLength" : 10000,
            "minLength" : 0
          },
          "subcategoryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          },
          "urgency" : {
            "type" : "string",
            "enum" : [ "LOW", "MEDIUM", "HIGH" ]
          }
        },
        "required" : [ "categoryId", "description", "title" ]
      },
...
      "FieldErrorDetail" : {
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "field", "message" ]
      },
...
      "IncidentDetails" : {
        "type" : "object",
        "properties" : {
          "affectedService" : {
            "type" : [ "string", "null" ]
          },
          "reopenCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "resolutionCode" : {
            "type" : [ "string", "null" ]
          },
          "resolutionNotes" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "affectedService", "reopenCount", "resolutionCode", "resolutionNotes" ]
      },
...
      "NamedRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "name" ]
      },
      "PageResponseTicketSummary" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TicketSummary"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "content", "page", "size", "totalElements", "totalPages" ]
      },
...
      "ServiceRequestDetails" : {
        "type" : "object",
        "properties" : {
          "catalogueItemId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "fulfilmentNotes" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "catalogueItemId", "fulfilmentNotes" ]
      },
...
      "TicketCreatedResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "priority" : {
            "type" : "string",
            "enum" : [ "P1", "P2", "P3", "P4" ]
          },
          "reference" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "INCIDENT", "SERVICE_REQUEST" ]
          }
        },
        "required" : [ "id", "priority", "reference", "status", "type" ]
      },
      "TicketResponse" : {
        "type" : "object",
        "properties" : {
          "assignedTeam" : {
            "$ref" : "#/components/schemas/NamedRef"
          },
          "assignee" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/NamedRef"
            }, {
              "type" : "null"
            } ]
          },
          "category" : {
            "$ref" : "#/components/schemas/NamedRef"
          },
          "closedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "description" : {
            "type" : "string"
          },
          "firstRespondedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "impact" : {
            "type" : "string",
            "enum" : [ "LOW", "MEDIUM", "HIGH" ]
          },
          "incident" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/IncidentDetails"
            }, {
              "type" : "null"
            } ]
          },
          "priority" : {
            "type" : "string",
            "enum" : [ "P1", "P2", "P3", "P4" ]
          },
          "reference" : {
            "type" : "string"
          },
          "requester" : {
            "$ref" : "#/components/schemas/NamedRef"
          },
          "resolvedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "serviceRequest" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/ServiceRequestDetails"
            }, {
              "type" : "null"
            } ]
          },
          "status" : {
            "type" : "string"
          },
          "subcategory" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/NamedRef"
            }, {
              "type" : "null"
            } ]
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "INCIDENT", "SERVICE_REQUEST" ]
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "urgency" : {
            "type" : "string",
            "enum" : [ "LOW", "MEDIUM", "HIGH" ]
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "assignedTeam", "assignee", "category", "closedAt", "createdAt", "description", "firstRespondedAt", "id", "impact", "incident", "priority", "reference", "requester", "resolvedAt", "serviceRequest", "status", "subcategory", "title", "type", "updatedAt", "urgency", "version" ]
      },
      "TicketSummary" : {
        "type" : "object",
        "properties" : {
          "assignedTeam" : {
            "$ref" : "#/components/schemas/NamedRef"
          },
          "assignee" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/NamedRef"
            }, {
              "type" : "null"
            } ]
          },
          "category" : {
            "$ref" : "#/components/schemas/NamedRef"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "firstRespondedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "priority" : {
            "type" : "string",
            "enum" : [ "P1", "P2", "P3", "P4" ]
          },
          "reference" : {
            "type" : "string"
          },
          "requester" : {
            "$ref" : "#/components/schemas/NamedRef"
          },
          "resolvedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "status" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "INCIDENT", "SERVICE_REQUEST" ]
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "assignedTeam", "assignee", "category", "createdAt", "firstRespondedAt", "id", "priority", "reference", "requester", "resolvedAt", "status", "title", "type", "updatedAt" ]
      },
...
      "TransitionRequest" : {
        "type" : "object",
        "properties" : {
          "notes" : {
            "type" : "string",
            "maxLength" : 10000,
            "minLength" : 0
          },
          "reason" : {
            "type" : "string",
            "maxLength" : 1000,
            "minLength" : 0
          },
          "resolutionCode" : {
            "type" : "string",
            "maxLength" : 40,
            "minLength" : 0
          },
          "targetStatus" : {
            "type" : "string",
            "minLength" : 1
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "targetStatus", "version" ]
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from POST /api/auth/login",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}